pyproject.toml
[project]
name = "vector-db-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "cohere>=4.37",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0"
]
